updated dto
